selenium/requirements.txt
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.0
idna==2.9
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.1
python-dotenv==0.12.0
requests==2.23.0
selenium==3.141.0
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.9
webdriver-manager==2.3.0
